remove TaskConfiguration by inlining all methods
