+++ subprojects/build-init/src/test/groovy/org/gradle/buildinit/plugins/WrapperPluginSpec.groovy
@@ -18,7 +18,6 @@
 package org.gradle.buildinit.plugins
 
 import org.gradle.api.tasks.wrapper.Wrapper
-import org.gradle.buildinit.tasks.internal.TaskConfiguration
 import org.gradle.test.fixtures.file.TestNameTestDirectoryProvider
 import org.gradle.util.TestUtil
 import org.gradle.util.UsesNativeServices
@@ -37,6 +36,6 @@ class WrapperPluginSpec extends Specification {
 
         then:
         project.tasks.wrapper instanceof Wrapper
-        project.tasks.wrapper.group == TaskConfiguration.GROUP
+        project.tasks.wrapper.group == "Build Setup"
     }
 }
